feat: support for amount bigint

+++ lib/src/algorand.dart
@@ -459,7 +459,7 @@ class Algorand {
   ///
   /// Throws an [AlgorandException] if there is an HTTP error.
   /// Returns the account's balance in microAlgos.
-  Future<int> getBalance(String address) async {
+  Future<BigInt> getBalance(String address) async {
     final response = await _indexer.getAccountByAddress(address);
 
     return response.account.amountWithoutPendingRewards;
